+++ src/opencntx/__init__.py
@@ -1,3 +1,3 @@
 """OPENCNTX public package."""
 
-__version__ = "1.7.3"
+__version__ = "1.7.4"